move old models to bmds2
