+++ bmds/bmds2/models/continuous.py
@@ -3,7 +3,7 @@ from copy import deepcopy
 
 import numpy as np
 
-from .. import constants
+from ... import constants
 from .base import BMDModel, DefaultParams
